canary: include simulator by default, make test programs build on it
canary_platforms() now defaults include_sim=True so every canary SConstruct
exercises a host-simulator build alongside the embedded targets. The twi
canary opts out (twi_tests sits on an F3-specific driver path; not worth
the ergonomic cost of a sim branch in twi.c for canary coverage).

test-pin.h now owns the cross-platform gpio abstraction: it pulls in
core/hardware.h on real targets and provides no-op gpio_* stubs plus a
placeholder TEST_PIN under SIM. Test programs drop the bare
#include "core/hardware.h" and just include "../test-pin.h" — or drop it
entirely if they don't touch GPIO (charecho, lineecho, linecounter,
delaytest, recv).

Incidental fixes uncovered by building on the 64-bit host:
  - charecho: %d → %zu for a size_t argument
  - timertest: cast void*<->int via intptr_t so 64-bit sim doesn't
    warn about pointer/integer size mismatches

+++ src/app/tests/timertest/timertest.c
@@ -16,7 +16,6 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-#include "core/hardware.h"
 #include "core/rulos.h"
 
 #define FREQ_USEC 500000
@@ -38,8 +37,8 @@ static void test_func(void *data) {
   gpio_set(TEST_PIN);
   schedule_us(FREQ_USEC * NUM_TASKS, test_func, data);
   gpio_clr(TEST_PIN);
-  LOG("line %d printed by task %d at time %" PRId32, i++, (int)data,
-      clock_time_us());
+  LOG("line %d printed by task %d at time %" PRId32, i++,
+      (int)(intptr_t)data, clock_time_us());
 }
 #endif
 
@@ -70,10 +69,10 @@ int main() {
   init_clock(FREQ_USEC, TIMER1);
 
   for (int i = 0; i < NUM_TASKS; i++) {
-    schedule_us(FREQ_USEC * (i + 1), test_func, (void *)i);
+    schedule_us(FREQ_USEC * (i + 1), test_func, (void *)(intptr_t)i);
   }
 
-  schedule_now(test_func, (void *)-1);
+  schedule_now(test_func, (void *)(intptr_t)-1);
 
   scheduler_run();
 #endif
